+++ RELEASE_NOTES.md
@@ -12,7 +12,9 @@ _Date TBD_
 * Add certificationDateStart and certificationDateEnd as advanced search parameters
 * Change corrective action plan search parameters to new surveillance search parameters - NOT BACKWARDS COMPATIBLE
 * Send questionable activity email when product owner changes (not during a merge)
-* Add /surveillance/download[?all=true] endpoints to allow download of CSV file with all surveillance and download of CSV file with surveillance that resulted in nonconformities
+* Add /surveillance/download[?type=all|basic|nonconformities] endpoint to allow download of CSV file with all surveillance and download of CSV file with surveillance that resulted in nonconformities
+* Add lookup to new tables for certification status event history
+* Add date of last certification status change to certified product details
 
 ### Bugs Fixed
 * Changed transition of care calculations for the EHR Certification ID to more closely match the rule.